+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- `black`, `white`, `gray`, `red`, `green` and `blue` constructors to the `Color` struct.
+
 ### Removed
 - `failure` dependency.